use 64-bit integers for literals

+++ pijama-core/src/literal.rs
@@ -5,7 +5,7 @@ use pijama_ty::base::BaseTy;
 /// The bit representation of a literal follows Rust's data layout.
 #[derive(Debug, Clone, Copy)]
 pub struct Literal {
-    bits: isize,
+    bits: i64,
     ty: BaseTy,
 }
 
@@ -14,14 +14,15 @@ impl Literal {
     pub fn base_ty(&self) -> BaseTy {
         self.ty
     }
+
     /// The bit representation of the literal.
-    pub fn bits(&self) -> isize {
+    pub fn bits(&self) -> i64 {
         self.bits
     }
 }
 
-impl From<isize> for Literal {
-    fn from(int: isize) -> Self {
+impl From<i64> for Literal {
+    fn from(int: i64) -> Self {
         Self {
             bits: int,
             ty: BaseTy::Integer,
@@ -32,7 +33,7 @@ impl From<isize> for Literal {
 impl From<bool> for Literal {
     fn from(b: bool) -> Self {
         Self {
-            bits: b as isize,
+            bits: b as i64,
             ty: BaseTy::Bool,
         }
     }
